requirements.txt
streamlit
langchain
openai
PyPDF2
pandas
python-dotenv
langchain-openai
langchain-community